MOO.AST

Description

Data structures representing an abstract syntax tree for MOO code

Data Structures

newtype Program

Constructors

Program [Statement]

data Statement

Constructors

Expression !LineNo Expr

If !LineNo Expr Then [ElseIf] Else

ForList !LineNo Id Expr [Statement]

ForRange !LineNo Id (Expr, Expr) [Statement]

While !LineNo (Maybe Id) Expr [Statement]

Fork !LineNo (Maybe Id) Expr [Statement]

Break (Maybe Id)

Continue (Maybe Id)

Return !LineNo (Maybe Expr)

TryExcept [Statement] [Except]

TryFinally [Statement] Finally

newtype Then

Constructors

Then [Statement]

data ElseIf

Constructors

ElseIf !LineNo Expr [Statement]

newtype Else

Constructors

Else [Statement]

data Except

Constructors

Except !LineNo (Maybe Id) Codes [Statement]

newtype Finally

Constructors

Finally [Statement]

data Expr

Constructors

Literal Value

List [Argument]

Variable Id

PropertyRef Expr Expr

Assign Expr Expr

Scatter [ScatterItem] Expr

VerbCall Expr Expr [Argument]

BuiltinFunc Id [Argument]

Expr `Index` Expr

Expr `Range` (Expr, Expr)

Length

Expr `In` Expr

Expr `Plus` Expr

Expr `Minus` Expr

Expr `Times` Expr

Expr `Divide` Expr

Expr `Remain` Expr

Expr `Power` Expr

Negate Expr

Conditional Expr Expr Expr

Expr `And` Expr

Expr `Or` Expr

Not Expr

Expr `CompareEQ` Expr

Expr `CompareNE` Expr

Expr `CompareLT` Expr

Expr `CompareLE` Expr

Expr `CompareGT` Expr

Expr `CompareGE` Expr

Catch Expr Codes Default

data Codes

Constructors

ANY

Codes [Argument]

newtype Default

Constructors

Default (Maybe Expr)

data Argument

Constructors

ArgNormal Expr

ArgSplice Expr

data ScatterItem

Constructors

ScatRequired Id

ScatOptional Id (Maybe Expr)

ScatRest Id

Utility Functions

isLValue :: Expr -> Bool

Can the given expression be used on the left-hand side of an assignment?

precedence :: Expr -> Int

Return a precedence value for the given expression, needed by MOO.Unparser to determine whether parentheses are necessary to isolate an expression from its surrounding context.