MOO.Builtins.Common

Documentation

type BuiltinSpec = (Id, (Builtin, Info))

A pair associating a named built-in function with its implementation and a description of the required arguments

type Builtin = [Value] -> MOO Value

The type of each built-in function implementation

data Info

A description of the number and types of required and optional arguments

Info min max [types] return-type describes a function requiring at least min and at most max arguments (or no maximum if max is Nothing), with each argument having the specified Type. The return-type is for documentation purposes and is not otherwise currently used.

Constructors

Info Int (Maybe Int) [Type] Type

defaults :: [Value] -> [Value] -> [Value]

defaults optional [default-values] generates an argument list by supplying default-values for any missing from optional.

booleanDefaults :: [Value] -> [Bool] -> [Bool]

Generate a list of boolean arguments with the given defaults.

maybeDefaults :: [Value] -> [Maybe Value]

Generate an infinite list of arguments where each missing argument is Nothing.