MOO.Builtins.Match

Description

Regular expression matching via PCRE through the FFI

Documentation

data Regexp

data MatchResult

Constructors

MatchFailed

MatchAborted

MatchSucceeded [(Int, Int)]

Compiling

newRegexp :: Text -> Bool -> IO (Either (String, CInt) Regexp)

newRegexp regexp case-matters compiles a regular expression pattern regexp into a Regexp value, or returns an error description if the pattern is malformed. The returned CInt is a byte offset into an internally translated pattern, and thus is probably not very useful.

Matching

match :: Regexp -> Text -> IO MatchResult

rmatch :: Regexp -> Text -> IO MatchResult