MOO.Command

Description

MOO command parsing and execution

Data Structures

data Command

A structure describing a player's parsed command

Constructors

Command

Fields

commandVerb :: StrT

commandArgs :: [StrT]

commandArgStr :: StrT

commandDObjStr :: StrT

commandPrepSpec :: PrepSpec

commandPrepStr :: StrT

commandIObjStr :: StrT

Parsing Typed Commands

parseWords :: Text -> [StrT]

Split a typed command into words according to the MOO rules for quoting and escaping.

parseCommand :: Text -> Command

Return a Command value describing the arguments of a typed command as parsed into verb, preposition, direct and indirect object, etc.

Executing MOO Code

runCommand :: Command -> MOO Value

Execute a typed command by locating and calling an appropriate MOO verb for the current player, matching dobj and iobj objects against the strings in the typed command.