MOO.Compiler

Description

Compiling abstract syntax trees into MOO computations

Documentation

compile :: Program -> MOO Value

Compile a complete MOO program into a computation in the MOO monad that returns whatever the MOO program returns.

evaluate :: Expr -> MOO Value

Compile a MOO expression into a computation in the MOO monad. If a MOO exception is raised and the current verb frame's debug bit is not set, return the error code as a MOO value rather than propagating the exception.