MOO.Connection

Documentation

data Connection

type ConnectionHandler = ConnectionName -> (Producer ByteString IO (), Consumer ByteString IO ()) -> IO ()

connectionHandler :: TVar World -> ObjId -> Bool -> ConnectionHandler

firstConnectionId :: ObjId

The first un-logged-in connection ID. Avoid conflicts with #-1 ($nothing), -3 ($failed_match).

withConnections :: (Map ObjId Connection -> MOO a) -> MOO a

Do something with the Map of all active connections.

withConnection :: ObjId -> (Connection -> MOO a) -> MOO a

Do something with the connection for the given object, raising E_INVARG if no such connection exists.

withMaybeConnection :: ObjId -> (Maybe Connection -> MOO a) -> MOO a

Do something with the connection for given object, if any.

connectionName :: Connection -> ConnectionName

connectionConnectedTime :: Connection -> TVar (Maybe UTCTime)

connectionActivityTime :: Connection -> TVar UTCTime

connectionOutputDelimiters :: Connection -> TVar (Text, Text)

sendToConnection :: Connection -> Text -> STM ()

closeConnection :: Connection -> STM ()

Initiate the closing of a connection by closing its output queue.

notify :: ObjId -> StrT -> MOO Bool

Send data to a connection, flushing if necessary.

notify' :: Bool -> ObjId -> StrT -> MOO Bool

Send data to a connection, optionally without flushing.

bufferedOutputLength :: Maybe Connection -> STM Int

Return the number of items currently buffered for output to a connection, or the maximum number of items that will be buffered up for output on any connection.

forceInput :: Bool -> ObjId -> StrT -> MOO ()

Force a line of input for a connection.

flushInput :: Bool -> Connection -> STM ()

Flush a connection's input queue, optionally showing what was flushed.

bootPlayer :: ObjId -> MOO ()

Close the connection associated with an object.

bootPlayer' :: Bool -> ObjId -> MOO ()

Close the connection associated with an object, with message varying on whether the object is being recycled.

setConnectionOption :: ObjId -> Id -> Value -> MOO ()

Set a connection option for an active connection.

getConnectionOptions :: ObjId -> MOO (Map Id Value)

Return a Map of all currently set connection options for a connection.