MOO.Database

Documentation

data Database

type Connected = [(ObjId, ObjId)]

data ServerOptions

Constructors

Options

Fields

bgSeconds :: Int

The number of seconds allotted to background tasks

bgTicks :: Int

The number of ticks allotted to background tasks

connectTimeout :: IntT

The maximum number of seconds to allow an un-logged-in in-bound connection to remain open

defaultFlushCommand :: Text

The initial setting of each new connection’s flush command

fgSeconds :: Int

The number of seconds allotted to foreground tasks

fgTicks :: Int

The number of ticks allotted to foreground tasks

maxStackDepth :: Int

The maximum number of levels of nested verb calls

queuedTaskLimit :: Maybe Int

The default maximum number of tasks a player can have

nameLookupTimeout :: IntT

The maximum number of seconds to wait for a network hostname/address lookup

outboundConnectTimeout :: IntT

The maximum number of seconds to wait for an outbound network connection to successfully open

protectProperty :: Id -> Bool

Restrict reading of built-in property to wizards

protectFunction :: Id -> Bool

Restrict use of built-in function to wizards

supportNumericVerbnameStrings :: Bool

Enables use of an obsolete verb-naming mechanism

data Persistence

Constructors

Persistence

Fields

persistenceVSpace :: VSpace

persistenceVersion :: PVar VVersion

persistenceDatabase :: PVar Database

persistenceConnected :: PVar Connected

persistenceCheckpoint :: PVar VUTCTime

serverOptions :: Database -> ServerOptions

initDatabase :: Database

dbObjectRef :: ObjId -> Database -> Maybe (PVar (Maybe Object))

dbObject :: ObjId -> Database -> VTx (Maybe Object)

maxObject :: Database -> ObjId

resetMaxObject :: Database -> VTx Database

renumber :: ObjId -> Database -> VTx (ObjId, Database)

Renumber an object in the database to be the least nonnegative object number not currently in use pursuant to the renumber() built-in function.

setObjects :: VSpace -> [Maybe Object] -> Database -> IO Database

addObject :: Object -> Database -> VTx Database

deleteObject :: ObjId -> Database -> VTx ()

modifyObject :: ObjId -> Database -> (Object -> VTx Object) -> VTx ()

allPlayers :: Database -> [ObjId]

setPlayer :: Bool -> ObjId -> Database -> Database

getServerOption :: Id -> MOO (Maybe Value)

getServerOption' :: ObjId -> Id -> MOO (Maybe Value)

loadServerOptions :: MOO ()

getServerMessage :: ObjId -> Id -> MOO [Text] -> MOO [Text]

loadPersistence :: VCache -> IO Persistence

syncPersistence :: Persistence -> IO ()

saveDatabase :: VCache -> (Database, Connected) -> IO ()