MOO.List

Description

Abstract MOO list type

Documentation

data MOOList

Accessors

Length information

length :: MOOList -> Int

null :: MOOList -> Bool

Indexing

(!) :: MOOList -> Int -> Value

head :: MOOList -> Value

Extracting sublists (slicing)

slice :: Int -> Int -> MOOList -> MOOList

tail :: MOOList -> MOOList

splitAt :: Int -> MOOList -> (MOOList, MOOList)

Construction

Initialization

empty :: MOOList

Concatenation

snoc :: MOOList -> Value -> MOOList

concat :: [MOOList] -> MOOList

Elementwise operations

Monadic mapping

forM_ :: Monad m => MOOList -> (Value -> m b) -> m ()

Working with predicates

Searching

elem :: Value -> MOOList -> Bool

findIndex :: (Value -> Bool) -> MOOList -> Maybe Int

elemIndex :: Value -> MOOList -> Maybe Int

Folding

Monadic folds

foldM :: Monad m => (a -> Value -> m a) -> a -> MOOList -> m a

Conversions

Lists

toList :: MOOList -> [Value]

fromList :: [Value] -> MOOList

MOO primitives

storageBytes :: MOOList -> Int

equal :: MOOList -> MOOList -> Bool

Association list interface

assocLens :: StrT -> MOOList -> Maybe (Maybe Value, Maybe Value -> MOOList)

Return the current value (if any) associated with the given key, and a function to associate the key with a new value (or remove it). Returns Nothing if the list is not a proper association list.

Convenience functions

set :: MOOList -> Value -> Int -> MOOList

Return a modified list with the given 0-based index replaced with the given value.

insert :: MOOList -> Int -> Value -> MOOList

Return a modified list with the given value inserted at the given 0-based index.

delete :: MOOList -> Int -> MOOList

Return a modified list with the value at the given 0-based index removed.