MOO.Parser

Documentation

data Program

parse :: Text -> Either Errors Program

runParser :: Stream s Identity t => Parsec s u a -> u -> SourceName -> s -> Either ParseError a

The most general way to run a parser over the Identity monad. runParser p state filePath input runs parser p on the input list of tokens input, obtained from source filePath with the initial user state st. The filePath is only used in error messages and may be the empty string. Returns either a ParseError (Left) or a value of type a (Right).

  parseFromFile p fname
    = do{ input <- readFile fname
        ; return (runParser p () fname input)
        }

initParserState :: ParserState

expression :: MOOParser Expr

between :: Stream s m t => ParsecT s u m open -> ParsecT s u m close -> ParsecT s u m a -> ParsecT s u m a

between open close p parses open, followed by p and close. Returns the value returned by p.

  braces  = between (symbol "{") (symbol "}")

whiteSpace :: ParsecT Text u Identity ()

eof :: (Stream s m t, Show t) => ParsecT s u m ()

This parser only succeeds at the end of the input. This is not a primitive parser but it is defined using notFollowedBy.

  eof  = notFollowedBy anyToken <?> "end of input"

program :: MOOParser Program

parseInt :: Text -> Maybe Value

parseFlt :: Text -> Maybe Value

parseNum :: Text -> Maybe Value

parseObj :: Text -> Maybe Value

keywords :: [String]