MOO.Task

Documentation

type MOO = ReaderT Environment (ContT TaskDisposition (StateT TaskState STM))

This is the basic MOO monad transformer stack. A computation of type MOO a is an STM transaction that returns a value of type a within an environment that supports state, continuations, and local modification.

data World

The known universe, as far as the MOO server is concerned

Constructors

World

Fields

database :: Database

The database of objects

tasks :: Map TaskId Task

Queued and running tasks

listeners :: Map PortNumber Listener

Network listening points

connections :: Map ObjId Connection

Network connections

nextConnectionId :: ObjId

The (negative) object number to be assigned to the next inbound or outbound connection

data Task

A structure representing a queued or running task

Constructors

Task

Fields

taskId :: TaskId

taskStatus :: TaskStatus

taskThread :: ThreadId

taskWorld :: TVar World

taskPlayer :: ObjId

taskState :: TaskState

taskComputation :: MOO Value

data TaskStatus

The running state of a task

Constructors

Pending

Running

Forked

Suspended Wake

Reading

data TaskDisposition

The intermediate or final result of a running task

Constructors

Complete Value

Suspend (Maybe Integer) (Resume ())

Read ObjId (Resume Value)

forall a . RequestIO (IO a) (Resume a)

Uncaught Exception CallStack

Timeout Resource CallStack

Suicide

data Resource

Task resource limits

Constructors

Ticks

Seconds

newtype Wake

A function to call in order to wake a suspended task

Constructors

Wake (Value -> IO ())

newtype Resume a

A continuation to resume the execution of a task where it left off

Constructors

Resume (a -> MOO Value)

newtype DelayedIO

An IO computation to be performed after the current task commits its STM transaction

Constructors

DelayedIO

Fields

runDelayed :: IO ()

data Environment

A Reader environment for state that either doesn't change, or can be locally modified for subcomputations

Constructors

Env

Fields

task :: Task

interruptHandler :: InterruptHandler

exceptionHandler :: ExceptionHandler

indexLength :: MOO Int

data TaskState

A State structure for data that may normally change during computation

Constructors

State

Fields

ticksLeft :: Int

stack :: CallStack

startTime :: UTCTime

randomGen :: StdGen

delayedIO :: DelayedIO

newtype CallStack

The stack of verb and/or built-in function frames

Constructors

Stack [StackFrame]

newtype Continuation

A local continuation for loop constructs

Constructors

Continuation (Value -> MOO Value)

data StackFrame

The data tracked for each verb and/or built-in function call

Constructors

Frame

Fields

depthLeft :: Int

contextStack :: [Context]

variables :: Map Id Value

debugBit :: Bool

permissions :: ObjId

verbName :: StrT

verbFullName :: StrT

verbLocation :: ObjId

initialThis :: ObjId

initialPlayer :: ObjId

builtinFunc :: Bool

lineNumber :: IntT

data Exception

A MOO exception

Constructors

Exception Code Message Value

initWorld :: World

initTask :: Task

newTaskId :: World -> StdGen -> STM TaskId

Generate and return a (random) TaskId not currently in use by any existing task.

newTask :: TVar World -> ObjId -> MOO Value -> IO Task

Create a pending Task for the given computation on behalf of the given player. A new TaskId is reserved for the task and the task is added to the World. (The task will not actually run until passed to runTask.)

taskOwner :: Task -> ObjId

isQueued :: TaskStatus -> Bool

queuedTasks :: MOO [Task]

timeoutException :: Resource -> Exception

stepTask :: Task -> IO (TaskDisposition, Task)

runTask :: Task -> IO (Maybe Value)

Run a task in a new Haskell thread, returning either the value produced by the task, or Nothing if the task suspends or aborts before producing a value. If the task suspends, it may continue running after this function returns. After the task is finished, it is removed from the task queue.

forkTask :: TaskId -> Integer -> MOO Value -> MOO ()

Create and queue a task to run the given computation after the given microsecond delay. E_INVARG may be raised if the delay is out of acceptable range. (The given TaskId should have been reserved by a call to newTaskId.)

interrupt :: TaskDisposition -> MOO a

Commit the current task's transaction, and return to the task dispatcher with an interrupt request. The task dispatcher may resume execution of the task later if the request is one which supplies an appropriate continuation.

requestIO :: IO a -> MOO a

Interrupt the current task to perform the given IO computation, and return the result. Note this implies a commit of the task's STM transaction.

liftSTM :: STM a -> MOO a

Lift an STM transaction into the MOO monad.

initEnvironment :: Task -> Environment

initState :: TaskState

newState :: IO TaskState

getWorld :: MOO World

getWorld' :: MOO (TVar World)

putWorld :: World -> MOO ()

modifyWorld :: (World -> World) -> MOO ()

getTask :: TaskId -> MOO (Maybe Task)

putTask :: Task -> MOO ()

purgeTask :: Task -> MOO ()

getDatabase :: MOO Database

putDatabase :: Database -> MOO ()

getPlayer :: MOO ObjId

getObject :: ObjId -> MOO (Maybe Object)

getProperty :: Object -> StrT -> MOO Property

getVerb :: Object -> Value -> MOO Verb

findVerb :: (Verb -> Bool) -> StrT -> ObjId -> MOO (Maybe ObjId, Maybe Verb)

callCommandVerb :: ObjId -> (ObjId, Verb) -> ObjId -> Command -> (ObjId, ObjId) -> MOO Value

callVerb :: ObjId -> ObjId -> StrT -> [Value] -> MOO Value

callFromFunc :: Id -> IntT -> (ObjId, StrT) -> [Value] -> MOO (Maybe Value)

evalFromFunc :: Id -> IntT -> MOO Value -> MOO Value

runVerb :: Verb -> StackFrame -> MOO Value

runTick :: MOO ()

modifyProperty :: Object -> StrT -> (Property -> MOO Property) -> MOO ()

modifyVerb :: (ObjId, Object) -> Value -> (Verb -> MOO Verb) -> MOO ()

readProperty :: ObjId -> StrT -> MOO (Maybe Value)

writeProperty :: ObjId -> StrT -> Value -> MOO ()

setBuiltinProperty :: (ObjId, Object) -> StrT -> Value -> MOO ()

initFrame :: StackFrame

formatFrames :: Bool -> [StackFrame] -> Value

pushFrame :: StackFrame -> MOO ()

popFrame :: MOO ()

activeFrame :: Task -> StackFrame

frame :: (StackFrame -> a) -> MOO a

caller :: (StackFrame -> a) -> MOO (Maybe a)

modifyFrame :: (StackFrame -> StackFrame) -> MOO ()

setLineNumber :: Int -> MOO ()

pushTryFinallyContext :: MOO Value -> MOO ()

pushLoopContext :: Maybe Id -> Continuation -> MOO ()

setLoopContinue :: Continuation -> MOO ()

popContext :: MOO ()

breakLoop :: Maybe Id -> MOO Value

continueLoop :: Maybe Id -> MOO Value

mkVariables :: [(Id, Value)] -> Map Id Value

Create a variable block for a verb by overriding the default.

catchException :: MOO a -> (Exception -> CallStack -> MOO a) -> MOO a

Install a local exception handler for the duration of the passed computation.

passException :: Exception -> CallStack -> MOO a

Re-raise an exception to the next enclosing handler.

raiseException :: Exception -> MOO a

Abort execution of the current computation and call the closest enclosing exception handler.

notyet :: String -> MOO a

Placeholder for features not yet implemented

raise :: Error -> MOO a

Create and raise an exception for the given MOO error.

isWizard :: ObjId -> MOO Bool

Determine whether the given object has its wizard bit set.

checkFloat :: FltT -> MOO Value

Verify that the given floating point number is neither infinite nor NaN, raising E_FLOAT or E_INVARG respectively if so. Also, return the corresponding MOO value.

checkProgrammer :: MOO ()

Verify that the current task permissions have programmer privileges, raising E_PERM if not.

checkWizard :: MOO ()

Verify that the current task permissions have wizard privileges, raising E_PERM if not.

checkPermission :: ObjId -> MOO ()

Verify that the current task permissions either have wizard privileges or are the same as the given object, raising E_PERM if not.

checkValid :: ObjId -> MOO Object

Verify that the given object is valid, raising E_INVARG if not. Also, return the referenced object.

checkFertile :: ObjId -> MOO ()

Verify that the given object is fertile for the current task permissions, raising E_PERM if not.

binaryString :: StrT -> MOO ByteString

Translate a MOO binary string into a Haskell ByteString, raising E_INVARG if the MOO string is improperly formatted.

random :: Random a => (a, a) -> MOO a

Generate and return a pseudorandom number in the given range, modifying the local generator state.

newRandomGen :: MOO StdGen

Split the local random number generator state in two, updating the local state with one of them and returning the other.

formatTraceback :: Exception -> CallStack -> [Text]

Generate traceback lines for an exception, suitable for displaying to a user.

delayIO :: IO () -> MOO ()

Perform the given IO computation after the current task commits its STM transaction.

Since IO can't be performed within a transaction, this is a simple alternative when the value returned by the IO isn't needed.