MOO.Version

Documentation

serverVersion :: Version

The current version of the server code, as a Version value

serverVersionText :: Text

The current version of the server code, as a displayable Text value (for use by the server_version() built-in function)