MOO.AST

Description

Data structures representing an abstract syntax tree for MOO code

Data Structures

newtype Program

Constructors

Program [Statement]

data Statement

Constructors

Expression !Int Expr

If !Int Expr Then [ElseIf] Else

ForList !Int Id Expr [Statement]

ForRange !Int Id (Expr, Expr) [Statement]

While !Int (Maybe Id) Expr [Statement]

Fork !Int (Maybe Id) Expr [Statement]

Break (Maybe Id)

Continue (Maybe Id)

Return !Int (Maybe Expr)

TryExcept [Statement] [Except]

TryFinally [Statement] Finally

newtype Then

Constructors

Then [Statement]

data ElseIf

Constructors

ElseIf !Int Expr [Statement]

newtype Else

Constructors

Else [Statement]

data Except

Constructors

Except !Int (Maybe Id) Codes [Statement]

newtype Finally

Constructors

Finally [Statement]

data Expr

Constructors

Literal Value

List [Arg]

Variable Id

PropRef Expr Expr

Assign Expr Expr

ScatterAssign [ScatItem] Expr

VerbCall Expr Expr [Arg]

BuiltinFunc Id [Arg]

Expr Index Expr

Expr Range (Expr, Expr)

Length

Expr In Expr

Expr Plus Expr

Expr Minus Expr

Expr Times Expr

Expr Divide Expr

Expr Remain Expr

Expr Power Expr

Negate Expr

Conditional Expr Expr Expr

Expr And Expr

Expr Or Expr

Not Expr

Expr Equal Expr

Expr NotEqual Expr

Expr LessThan Expr

Expr LessEqual Expr

Expr GreaterThan Expr

Expr GreaterEqual Expr

Catch Expr Codes Default

data Codes

Constructors

ANY

Codes [Arg]

newtype Default

Constructors

Default (Maybe Expr)

data Arg

Constructors

ArgNormal Expr

ArgSplice Expr

data ScatItem

Constructors

ScatRequired Id

ScatOptional Id (Maybe Expr)

ScatRest Id

Utility Functions

isLValue :: Expr -> Bool

Can the given expression be used on the left-hand side of an assignment?

precedence :: Expr -> Int

Return a precedence value for the given expression, needed by MOO.Unparser to determine whether parentheses are necessary to isolate an expression from its surrounding context.