MOO.Builtins.Common

Documentation

data Builtin

A complete description of a builtin function, including its name, the number and types of required and optional arguments, the return type, and the implementing Haskell function

Constructors

Builtin

Fields

builtinName :: Id

builtinMinArgs :: Int

builtinMaxArgs :: Maybe Int

builtinArgTypes :: [Type]

builtinReturnType :: Type

builtinFunction :: [Value] -> MOO Value

defaults :: [Value] -> [Value] -> [Value]

defaults optional [default-values] generates an argument list by supplying default-values for any missing from optional.

booleanDefaults :: [Value] -> [Bool] -> [Bool]

Generate a list of boolean arguments with the given defaults.

maybeDefaults :: [Value] -> [Maybe Value]

Generate an infinite list of arguments where each missing argument is Nothing.