MOO.Builtins.Match

Description

Regular expression matching via PCRE through the FFI

Documentation

data Regexp

data MatchResult

Constructors

MatchFailed

MatchAborted

MatchSucceeded [(Int, Int)]

Compiling

newRegexp

Arguments

:: Text
pattern

-> Bool
case matters?

-> Either (String, Int) Regexp

Compile a regular expression pattern into a Regexp value, or return an error description if the pattern is malformed. The returned Int is a byte offset into an internally translated pattern, and thus is probably not very useful.

Matching

match :: Regexp -> Text -> MatchResult

Perform regular expression matching.

rmatch :: Regexp -> Text -> MatchResult

Perform regular expression matching, returning the rightmost match.