MOO.Task

Monad Interface

type MOO = ReaderT Environment (ContT TaskDisposition (StateT TaskState VTx))

This is the basic MOO monad transformer stack. A computation of type MOO a is a VTx transaction (layered on STM) that returns a value of type a within an environment that supports state, continuations, and local modification.

data Environment

A Reader environment for state that either doesn't change, or can be locally modified for subcomputations

Constructors

Env

Fields

task :: Task

interruptHandler :: InterruptHandler

exceptionHandler :: ExceptionHandler

indexLength :: MOO Value

initEnvironment :: Task -> Environment

liftVTx :: VTx a -> MOO a

Lift a VTx transaction into the MOO monad.

liftSTM :: STM a -> MOO a

Lift an STM transaction into the MOO monad.

World Interface

data World

The known universe, as far as the MOO server is concerned

Constructors

World

Fields

writeLog :: Text -> STM ()

Logging function

persistence :: Persistence

Persistent storage

checkpoint :: STM ()

Database checkpoint signal

database :: Database

The database of objects

tasks :: Map TaskId Task

Queued and running tasks

listeners :: Map Point Listener

Network listening points

connections :: Map ObjId Connection

Network connections

nextConnectionId :: ObjId

The (negative) object number to be assigned to the next inbound or outbound connection

outboundNetwork :: Bool

Is open_network_connection() enabled?

bindAddress :: Maybe HostName

Interface address to bind to for incoming connections

shutdownMessage :: MVar Text

Shutdown signal

newWorld :: (Text -> STM ()) -> Persistence -> Bool -> IO (TVar World)

getWorld :: MOO World

getWorld' :: MOO (TVar World)

putWorld :: World -> MOO ()

modifyWorld :: (World -> World) -> MOO ()

updateConnections :: TVar World -> (Map ObjId Connection -> Map ObjId Connection) -> VTx ()

getDatabase :: MOO Database

putDatabase :: Database -> MOO ()

getVSpace :: MOO VSpace

serverOption :: (ServerOptions -> a) -> MOO a

Fetch the current setting of a server option obtained from $server_options.

Task Interface

data Task

A structure representing a queued or running task

Constructors

Task

Fields

taskId :: TaskId

taskStatus :: TaskStatus

taskThread :: ThreadId

taskWorld :: TVar World

taskPlayer :: ObjId

taskState :: TaskState

taskComputation :: MOO Value

data TaskStatus

The running state of a task

Constructors

Pending

Running

Forked

Suspended Wake

Reading

newtype Wake

A function to call in order to wake a suspended task

Constructors

Wake (Value -> IO ())

data TaskState

A State structure for data that may normally change during computation

Constructors

State

Fields

ticksLeft :: Int

secondsLimit :: Int

stack :: CallStack

startTime :: UTCTime

randomGen :: StdGen

delayedIO :: DelayedIO

newtype CallStack

The stack of verb and/or built-in function frames

Constructors

Stack [StackFrame]

newtype DelayedIO

An IO computation to be performed after the current task commits its STM transaction

Constructors

DelayedIO

Fields

runDelayed :: IO ()

data TaskDisposition

The intermediate or final result of a running task

Constructors

Complete Value

Suspend (Resume ())

forall a . RequestIO (IO a) (Resume a)

Uncaught Exception

Timeout Resource CallStack

Suicide

newtype Resume a

A continuation to resume the execution of a task where it left off

Constructors

Resume (a -> MOO Value)

data Resource

Task resource limits

Constructors

Ticks

Seconds

initState :: TaskState

newState :: IO TaskState

initTask :: Task

newTaskId :: World -> StdGen -> TaskId

Generate a (random) TaskId not currently in use by any existing task.

newTask :: TVar World -> ObjId -> MOO Value -> IO Task

Create a pending Task for the given computation on behalf of the given player. A new TaskId is reserved for the task and the task is added to the World. (The task will not actually run until passed to runTask.)

defaultMaxStackDepth :: Num a => a

defaultFgTicks :: Num a => a

defaultBgTicks :: Num a => a

defaultFgSeconds :: Num a => a

defaultBgSeconds :: Num a => a

getDelay :: Value -> MOO Integer

resetLimits :: Bool -> MOO ()

Reset the number of ticks and seconds available for the current task based on the latest values obtained from $server_options.

taskOwner :: Task -> ObjId

isQueued :: TaskStatus -> Bool

queuedTasks :: MOO [Task]

stepTask :: Task -> IO (TaskDisposition, Task)

runTask :: Task -> IO (Maybe Value)

Run a task in a new Haskell thread, returning either the value produced by the task, or Nothing if the task suspends or aborts before producing a value. If the task suspends, it may continue running after this function returns. The task is removed from the task queue after it is finished.

forkTask :: TaskId -> Integer -> MOO Value -> MOO ()

Create and queue a task to run the given computation after the given microsecond delay. E_INVARG may be raised if the delay is out of acceptable range. (The given TaskId should have been reserved by a call to newTaskId.)

interrupt :: TaskDisposition -> MOO a

Commit the current task's transaction, and return to the task dispatcher with an interrupt request. The task dispatcher may resume execution of the task later if the request is one which supplies an appropriate continuation.

requestIO :: IO a -> MOO a

Interrupt the current task to perform the given IO computation, and return the result. Note this implies a commit of the task's STM transaction.

delayIO :: IO () -> MOO ()

Perform the given IO computation after the current task commits its STM transaction.

Since general IO can't be performed within a transaction, this is a simple alternative when the value returned by the IO isn't needed.

unsafeIOtoMOO :: IO a -> MOO a

A version of catchUnsafeIOtoMOO that simply propagates any thrown exception into the calling thread, most likely aborting its execution.

catchUnsafeIOtoMOO :: IO a -> (SomeException -> MOO a) -> MOO a

Unsafely perform the given IO action within the current STM transaction, using the supplied exception handler in case the IO throws an exception.

Since STM transactions may be aborted at any time, the IO is performed in a separate thread in order to guarantee consistency with any finalizers, brackets, and so forth. The IO must be idempotent as it may be run more than once.

Note that all the hazards of unsafePerformIO apply; in particular, it is incumbent upon the caller to ensure the IO action is executed at least as many times as desired (and not, say, optimized to a single execution).

getTask :: TaskId -> MOO (Maybe Task)

putTask :: Task -> MOO ()

purgeTask :: Task -> MOO ()

Object Interface

getPlayer :: MOO ObjId

getObject :: ObjId -> MOO (Maybe Object)

getObjectName :: ObjId -> MOO StrT

getProperty :: Object -> StrT -> MOO Property

modifyProperty :: Object -> StrT -> (Property -> MOO Property) -> MOO ()

modifyVerb :: (ObjId, Object) -> Value -> (Verb -> MOO Verb) -> MOO ()

readProperty :: ObjId -> StrT -> MOO (Maybe Value)

writeProperty :: ObjId -> StrT -> Value -> MOO ()

setBuiltinProperty :: (ObjId, Object) -> StrT -> Value -> MOO ()

Verb Execution Interface

getVerb :: Object -> Value -> MOO Verb

findVerb :: (Verb -> Bool) -> StrT -> ObjId -> MOO (Maybe ObjId, Maybe Verb)

callSystemVerb :: StrT -> [Value] -> MOO (Maybe Value)

callSystemVerb' :: ObjId -> StrT -> [Value] -> StrT -> MOO (Maybe Value)

callCommandVerb :: ObjId -> (ObjId, Verb) -> ObjId -> Command -> ObjId -> ObjId -> MOO Value

callVerb :: ObjId -> ObjId -> StrT -> [Value] -> MOO Value

callFromFunc :: StrT -> LineNo -> (ObjId, StrT) -> [Value] -> MOO (Maybe Value)

evalFromFunc :: StrT -> LineNo -> MOO Value -> MOO Value

runVerb :: Verb -> StackFrame -> MOO Value

runTick :: MOO ()

Verb Frame Interface

data StackFrame

The data tracked for each verb and/or built-in function call

Constructors

Frame

Fields

depthLeft :: Int

contextStack :: [Context]

variables :: HashMap Id Value

debugBit :: Bool

permissions :: ObjId

verbName :: StrT

verbFullName :: StrT

verbLocation :: ObjId

initialThis :: ObjId

initialPlayer :: ObjId

builtinFunc :: Bool

lineNumber :: LineNo

newtype Continuation

A local continuation for loop constructs

Constructors

Continuation (() -> MOO Value)

initFrame :: StackFrame

formatFrames :: Bool -> [StackFrame] -> Value

activeFrame :: Task -> StackFrame

frame :: (StackFrame -> a) -> MOO a

caller :: (StackFrame -> a) -> MOO (Maybe a)

modifyFrame :: (StackFrame -> StackFrame) -> MOO ()

setLineNumber :: LineNo -> MOO ()

mkVariables :: [(Id, Value)] -> HashMap Id Value

Create a variable block for a verb by overriding the default.

Loop and Try/Finally Control Functions

pushTryFinallyContext :: MOO Value -> MOO ()

pushLoopContext :: Maybe Id -> Continuation -> MOO ()

setLoopContinue :: Continuation -> MOO ()

popContext :: MOO ()

breakLoop :: Maybe Id -> MOO Value

continueLoop :: Maybe Id -> MOO Value

Exception Handling

data Exception

A MOO exception

Constructors

Exception

Fields

exceptionCode :: Code

exceptionMessage :: Message

exceptionValue :: Value

exceptionCallStack :: CallStack

exceptionDebugBit :: Bool

A copy of the debug bit from the verb frame in which the exception was raised

type Code = Value

type Message = StrT

raiseException :: Code -> Message -> Value -> MOO a

Abort execution of the current computation and call the nearest enclosing exception handler.

raise :: Error -> MOO a

Create and raise an exception for the given MOO error.

catchException :: MOO a -> (Exception -> MOO a) -> MOO a

Install a local exception handler for the duration of the passed computation.

passException :: Exception -> MOO a

Re-raise an exception to the next enclosing handler.

handleDebug :: MOO Value -> MOO Value

Execute the passed computation, capturing any exception raised in verb frames with debug bit unset and returning the error code as an ordinary value instead of propagating the exception.

timeoutException :: Resource -> CallStack -> Exception

Utility Check Functions

isWizard :: ObjId -> MOO Bool

Determine whether the given object has its wizard bit set.

checkFloat :: FltT -> MOO Value

Verify that the given floating point number is neither infinite nor NaN, raising E_FLOAT or E_INVARG respectively if so. Also, return the corresponding MOO value.

checkProgrammer :: MOO ()

Verify that the current task permissions have programmer privileges, raising E_PERM if not.

checkWizard :: MOO ()

Verify that the current task permissions have wizard privileges, raising E_PERM if not.

checkPermission :: ObjId -> MOO ()

Verify that the current task permissions either have wizard privileges or are the same as the given object, raising E_PERM if not.

checkValid :: ObjId -> MOO Object

Verify that the given object is valid, raising E_INVARG if not. Also, return the referenced object.

checkFertile :: ObjId -> MOO ()

Verify that the given object is fertile for the current task permissions, raising E_PERM if not.

checkProtectedProperty :: Id -> MOO ()

Verify that the named built-in property is not protected by $server_options.protect_prop, or that the current task permissions have wizard privileges if it is, raising E_PERM otherwise.

checkRecurrence

Arguments

:: (Object -> Maybe ObjId)
relationship projection

-> ObjId
subject

-> ObjId
object to check

-> MOO ()

Verify that the given object does not have a recursive relationship with the given subject, raising E_RECMOVE if so.

checkQueuedTaskLimit :: MOO ()

Verify that the programmer has not reached their queued task limit (before creating a new forked, suspended, or reading task).

Miscellaneous

binaryString :: StrT -> MOO ByteString

Translate a MOO binary string into a Haskell ByteString, raising E_INVARG if the MOO string is improperly formatted.

random :: Random a => (a, a) -> MOO a

Generate and return a pseudorandom value in the given range, modifying the local generator state.

newRandomGen :: MOO StdGen

Split the local random number generator state in two, updating the local state with one of them and returning the other.

delay :: Integer -> IO ()

Wait for the given number of microseconds to elapse.

shutdown :: StrT -> MOO ()

Begin the server shutdown process.

notyet :: StrT -> MOO a

Placeholder for features not yet implemented