MOO.Types

Description

Basic data types used throughout the MOO server code

Haskell Types Representing MOO Values

type IntT

Arguments

= Int32
MOO integer

type FltT

Arguments

= Double
MOO floating-point number

type StrT

Arguments

= Text
MOO string

type ObjT

Arguments

= Int
MOO object number

type ErrT

Arguments

= Error
MOO error

type LstT

Arguments

= Vector Value
MOO list

type ObjId

Arguments

= ObjT
MOO object number

type Id

Arguments

= StrT
MOO identifier (string)

MOO Type and Value Reification

data Type

A Type represents one or more MOO value types.

Constructors

TAny
any type

TNum
integer or floating-point number

TInt
integer

TFlt
floating-point number

TStr
string

TObj
object number

TErr
error

TLst
list

data Value

A Value represents any MOO value.

Constructors

Int !IntT
integer

Flt !FltT
floating-point number

Str !StrT
string

Obj !ObjT
object number

Err !ErrT
error

Lst !LstT
list

data Error

A MOO error

Constructors

E_NONE
No error

E_TYPE
Type mismatch

E_DIV
Division by zero

E_PERM
Permission denied

E_PROPNF
Property not found

E_VERBNF
Verb not found

E_VARNF
Variable not found

E_INVIND
Invalid indirection

E_RECMOVE
Recursive move

E_MAXREC
Too many verb calls

E_RANGE
Range error

E_ARGS
Incorrect number of arguments

E_NACC
Move refused by destination

E_INVARG
Invalid argument

E_QUOTA
Resource limit exceeded

E_FLOAT
Floating-point arithmetic error

nothing :: Value

The default false value (zero)

Type and Value Functions

fromInt :: Value -> IntT

fromFlt :: Value -> FltT

fromStr :: Value -> StrT

fromObj :: Value -> ObjT

fromErr :: Value -> ErrT

fromLst :: Value -> LstT

equal :: Value -> Value -> Bool

Test two MOO values for indistinguishable (case-sensitive) equality.

truthOf :: Value -> Bool

Is the given MOO value considered to be true or false?

truthValue :: Bool -> Value

Return a default MOO value (integer) having the given boolean value.

typeOf :: Value -> Type

Return a Type indicating the type of the given value.

typeCode :: Type -> IntT

Return an integer code corresponding to the given type. These codes are visible to MOO code via the typeof() built-in function and various predefined variables.

toText :: Value -> Text

Return a string representation of the given MOO value, using the same rules as the tostr() built-in function.

toLiteral :: Value -> Text

Return a literal representation of the given MOO value, using the same rules as the toliteral() built-in function.

error2text :: Error -> Text

Return a string description of the given error value.

text2binary :: Text -> Maybe [Word8]

Parse a MOO binary string and return the corresponding byte values, or Nothing if the string is improperly formatted.

List Convenience Functions

fromList :: [Value] -> Value

fromListBy :: (a -> Value) -> [a] -> Value

stringList :: [StrT] -> Value

objectList :: [ObjT] -> Value

listSet :: LstT -> Int -> Value -> LstT

Miscellaneous

validStrChar :: Char -> Bool

May the given character appear in a valid MOO string?

endOfTime :: UTCTime

This is the last UTC time value representable as a 32-bit seconds-since-1970 value. Unfortunately it is used as a sentinel value in LambdaMOO to represent the starting time of indefinitely suspended tasks, so we really can't support time values beyond this point... yet.

Estimating Haskell Storage Sizes

class Sizeable t where

The Sizeable class is used to estimate the storage requirements of various values, for use by several built-in functions which are supposed to return the byte sizes of certain internal structures.

The sizes calculated by instances of this class are necessarily approximate, since it may be difficult or impossible to measure them precisely in the Haskell runtime environment.

Methods

storageBytes :: t -> Int

Return the estimated storage size of the given value, in bytes.